Fix intercal
Since thing related to syslib (such as DO (1000) NEXT where upto (1999))
went so weird I suffered from resolving the issue. Then got to `ick -eE
SOURCE.i syslibc`. Also when I was `make`ing on Dockerfile to build the
INTERCAL compiler I found "configure" file seemed to generate the
Makefile, I think, incorrectly such as `... ``(library name) ...`
because library name is not a command. So the TODO is to patch to such
something.